feat: conexión BLE completa y refactor estructural de BluetoothScreen
