+++ App.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import BluetoothScreen from './src/screens/BluetoothScreen';
+import BleScreen from './src/screens/BleScreen';
 import WiFiConfigScreen from './src/screens/WifiConfigScreen';
 import type { RootStackParamList } from './src/navigation/types';
 import PatientsScreen
@@ -14,15 +14,15 @@ import AuthForm from './src/views/loginView';
 import TutorProfile from './src/shared/components/profiles/TutorProfile'
 import TherapistProfile from './src/shared/components/profiles/TherapistProfile';
 import HomeTutor from './src/screens/tutor/HomeTutorScreen';
-import BluetoothScreen1 from './src/screens/BluetoothScreen1';
+import BluetoothScreen from './src/screens/BluetoothScreen';
 import WifiScreen1 from './src/screens/WifiScreen1';
 
 
 export default function App() {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Bluetooth">
-        <Stack.Screen name="Bluetooth" component={BluetoothScreen} options={{ title: 'Buscar Peluche' }} />
+      <Stack.Navigator initialRouteName="BleScreen">
+        <Stack.Screen name="BleScreen" component={BleScreen} options={{ title: 'LISTA DE PANTALLAS' }} />
         <Stack.Screen name="WiFi" component={WiFiConfigScreen} options={{ title: 'Conectar a WiFi' }} />
         <Stack.Screen name="Patients" component={PatientsScreen} options={{headerShown: false}} />
         <Stack.Screen name="TutorProfile" component={TutorProfile} options={{ headerShown: false}} />
@@ -31,7 +31,7 @@ export default function App() {
         <Stack.Screen name="Profile" component={ProfileScreen} options={{headerShown: false}} />
         <Stack.Screen name="Dashboard" component={DashboardScreen} options={{ title: 'Dashboard' }} />
         <Stack.Screen name="HomeTutor" component={HomeTutor} options={{headerShown: false}} />
-        <Stack.Screen name="Bluetooth1" component={BluetoothScreen1} options={{ headerShown: false }} />
+        <Stack.Screen name="Bluetooth" component={BluetoothScreen} options={{ headerShown: false }} />
         <Stack.Screen name="Wifi1" component={WifiScreen1} options={{ headerShown: false }} />
