+++ src/navigation/types.ts
@@ -2,7 +2,7 @@ import { Device } from 'react-native-ble-plx';
 
 export type RootStackParamList = {
   Dashboard: undefined;
-  Bluetooth: undefined;
+  BleScreen: undefined;
   WiFi: { device: Device };
   Patients: { openAddModal?: boolean };
   TutorProfile: undefined;
@@ -10,6 +10,6 @@ export type RootStackParamList = {
   Profile: undefined;
   TherapistProfile: undefined;
   HomeTutor:undefined;
-  Bluetooth1: undefined; // Nueva Ruta para BluetoothScreen1
+  Bluetooth: undefined; 
   Wifi1: { device: any }; // NUEVA RUTA para WifiScreen1, 'any' temporalmente si Device no es el mismo que en la simulación
 };
