+++ src/shared/assets/icons/OsitoIcon.tsx
