+++ src/shared/components/bluetooth/SimulatedDeviceScanner/SimulatedDeviceScanner.tsx
